Downloads finished; Index paging started for books, videos, articles

+++ TeachTheChild.Web/Areas/Moderator/Controllers/BooksController.cs
@@ -1,6 +1,7 @@
 ﻿namespace TeachTheChild.Web.Areas.Moderator.Controllers
 {
     using AutoMapper;
+    using Microsoft.AspNetCore.Hosting;
     using Microsoft.AspNetCore.Identity;
     using Microsoft.AspNetCore.Mvc;
     using System.Threading.Tasks;
@@ -9,6 +10,8 @@
     using TeachTheChild.Services.Moderator.Contracts;
     using TeachTheChild.Services.Moderator.Models.Books;
     using TeachTheChild.Web.Areas.Moderator.Models.Books;
+    using TeachTheChild.Web.Infrastructure.Constants;
+    using TeachTheChild.Web.Infrastructure.Helpers;
     using TeachTheChild.Web.Models;
 
     public class BooksController : BaseModeratorControler
@@ -17,17 +20,28 @@
         private readonly IUsersService usersService;
         private readonly UserManager<User> userManager;
         private readonly IMapper mapper;
+        private readonly IHostingEnvironment hostingEnvironment;
 
         public BooksController(
             IBooksModeratorService booksService, 
             UserManager<User> userManager,
             IUsersService usersService,
-            IMapper mapper)
+            IMapper mapper,
+            IHostingEnvironment hostingEnvironment)
         {
             this.booksService = booksService;
             this.userManager = userManager;
             this.usersService = usersService;
             this.mapper = mapper;
+            this.hostingEnvironment = hostingEnvironment;
+        }
+
+        private string RootPath
+        {
+            get
+            {
+                return this.hostingEnvironment.WebRootPath;
+            }
         }
 
         public IActionResult Index()
@@ -67,8 +81,19 @@
         [HttpPost]
         public async Task<IActionResult> Delete([FromBody]DeleteBookBindingModel model)
         {
-            // delete picture
-            var result = await this.booksService.DeleteAsync(model.Id);
+            var pictureUrl = await this.booksService.GetPictureUrlByIdAsync(model.Id);
+            if (pictureUrl == null)
+            {
+                return this.BadRequest(WebConstants.DeleteDownloadPictureEror);
+            }
+
+            var result = FileHelpers.Delete($"{this.RootPath}{pictureUrl}");
+            if (!result)
+            {
+                return this.BadRequest(WebConstants.DeleteDownloadPictureEror);
+            }
+
+            result = await this.booksService.DeleteAsync(model.Id);
 
             return this.Json(new { success = result });
         }
